+++ js/data.js
@@ -1,5 +1,3 @@
-import {getRandomInteger, getRandomArrayElement, createIdGenerator} from '.util.js';
-
 const PICTURE_COUNT = 25;
 const AVATAR_COUNT = 6;
 const LIKE_MIN_COUNT = 15;
@@ -36,37 +34,4 @@ const NAME = [
   'Марьяна'
 ];
 
-const generateCommentId = createIdGenerator();
-
-const createMessage = () =>
-  Array.from({
-      length: getRandomInteger(1, 2)
-    }, () =>
-    getRandomArrayElement(COMMENT_LINES)
-  ).join(' ');
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: 'img/avatar-${getRandomInteger(1, AVATAR_COUNT).svg',
-  message: createMessage(),
-  name: getRandomArrayElement(NAMES),
-});
-
-const createPicture = (index) => ({
-  id: index,
-  url: 'photos/$[index].jpg',
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(LIKE_MIN_COUNT, LIKE_MAX_COUNT),
-  comments: Array.from({
-      length: getRandomInteger(0, COMMENT_COUNT)
-    },
-    createComment
-  )
-});
-
-const getPictures = () =>
-  Array.from({ length: PICTURE_COUNT }, (_, pictureIndex) =>
-    createPicture(pictureIndex + 1)
-  );
-
-export {getPictures};
+export {PICTURE_COUNT, AVATAR_COUNT, LIKE_MIN_COUNT, LIKE_MAX_COUNT, COMMENT_COUNT, COMMENT_LINES, DESCRIPTIONS, NAME};
