+++ js/main.js
@@ -1,3 +1,3 @@
-import {getPictures} from './data.js'
+import {getPictures} from './generation.js'
 
 getPictures();
